move unused webpages into another folder
