+++ ai4good/models/cm/cm_model.py
@@ -1,4 +1,4 @@
-from abc import ABCMeta, abstractmethod
+from abc import abstractmethod
 
 from typeguard import typechecked
 
@@ -8,8 +8,6 @@ from ai4good.params.param_store import ParamStore
 from ai4good.models.cm.initialise_parameters import Parameters
 from ai4good.models.cm.simulator import Simulator, generate_csv
 from ai4good.utils.logger_util import get_logger
-from ai4good.webapp.cm_model_report_utils import normalize_report, prevalence_age_table, prevalence_all_table, \
-    cumulative_all_table, cumulative_age_table
 
 logger = get_logger(__name__)
