+++ ai4good/webapp/old_pages/cm_model_report_page.py
@@ -8,10 +8,9 @@ import dash_html_components as html
 import dash_core_components as dcc
 import dash_bootstrap_components as dbc
 
-from ai4good.models.cm.cm_model import CompartmentalModel
 from ai4good.models.cm.initialise_parameters import Parameters
 from ai4good.webapp.apps import dash_app, facade, model_runner, cache, local_cache, cache_timeout
-from ai4good.webapp.cm_model_report_utils import *
+from ai4good.webapp.old_pages.cm_model_report_utils import *
 from ai4good.webapp.metadata_report import GenerateMetadataDict, GenerateMetadataHTML
 from ai4good.utils.logger_util import get_logger
